+++ widgets/ListItemWidgets.py
@@ -19,7 +19,8 @@ email                : michaelsalgado@gkudos.com, info@gkudos.com
  ***************************************************************************/
 """
 
-from PyQt4.QtGui import QWidget
+from PyQt4.QtCore import Qt, QSize
+from PyQt4.QtGui import QWidget, QLabel
 
 from QgisCartoDB.ui.ListItem import Ui_ListItem
 
@@ -69,4 +70,14 @@ class CartoDBDatasetsListItem(QWidget):
 class CartoDBLayerListItem(CartoDBDatasetsListItem):
     def __init__(self, tableName=None, layer=None, size=None, rows=None):
         CartoDBDatasetsListItem.__init__(self, tableName, None, size, rows)
+
+        self.ui.statusLB = QLabel(self)
+        self.ui.statusLB.setMaximumSize(QSize(100, 16777215))
+        self.ui.statusLB.setAlignment(Qt.AlignCenter | Qt.AlignTrailing | Qt.AlignVCenter)
+        self.ui.statusLB.setWordWrap(True)
+        self.ui.horizontalLayout.insertWidget(1, self.ui.statusLB)
+
         self.layer = layer
+
+    def setStatus(self, status):
+        self.ui.statusLB.setText(status)